Add Apiresponse

+++ DioShop_Backend/DioShop.Application/Features/Brand/Handlers/Commands/DeleteBrandCommandHandler.cs
@@ -6,11 +6,12 @@ using System.Threading.Tasks;
 using AutoMapper;
 using DioShop.Application.Contracts.Infrastructure.IRepositories;
 using DioShop.Application.Features.Brand.Requests.Commands;
+using DioShop.Application.Ultils;
 using MediatR;
 
 namespace DioShop.Application.Features.Brand.Handlers.Commands
 {
-    public class DeleteBrandCommandHandler : IRequestHandler<DeleteBrandCommand,Unit>
+    public class DeleteBrandCommandHandler : IRequestHandler<DeleteBrandCommand, ApiResponse<object>>
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
@@ -21,17 +22,29 @@ namespace DioShop.Application.Features.Brand.Handlers.Commands
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(DeleteBrandCommand request, CancellationToken cancellationToken)
+        public async Task<ApiResponse<object>> Handle(DeleteBrandCommand request, CancellationToken cancellationToken)
         {
-
             var brand = await _unitOfWork.BrandRepository.Get(request.Id);
-
-
+            if (brand == null)
+            {
+                return new ApiResponse<object>
+                {
+                    Success = false,
+                    Message = $"Brand with ID {request.Id} does not exist",
+                    Data = null
+                };
+            }
 
             await _unitOfWork.BrandRepository.Delete(brand);
             await _unitOfWork.Save();
 
-            return Unit.Value;
+            return new ApiResponse<object>
+            {
+                Success = true,
+                Message = "Brand deleted successfully",
+                Data = null
+            };
         }
     }
+
 }
